# osmpbf: do not preallocate nodes