regarding https://github.com/paulmach/osm/issues/25 on high memory usage.

I noticed that pre-allocating the nodes could cause problems if you're saving some of them for later use. If just 1 node out of the whole 8000 node array is saved, the GC can not cleanup the any of them. 

Benchmarks are impacted:
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkLondon-12               235836166     248139745     +5.22%
BenchmarkLondon_nodes-12         163988036     177138150     +8.02%
BenchmarkLondon_ways-12          136805230     135422092     -1.01%
BenchmarkLondon_relations-12     82180166      81767765      -0.50%

benchmark                        old allocs     new allocs     delta
BenchmarkLondon-12               2416813        5145479        +112.90%
BenchmarkLondon_nodes-12         1003846        3732512        +271.82%
BenchmarkLondon_ways-12          1792714        1792718        +0.00%
BenchmarkLondon_relations-12     456772         456772         +0.00%

benchmark                        old bytes     new bytes     delta
BenchmarkLondon-12               954879689     952781068     -0.22%
BenchmarkLondon_nodes-12         649480586     647382889     -0.32%
BenchmarkLondon_ways-12          432818482     432819399     +0.00%
BenchmarkLondon_relations-12     179085842     179085459     -0.00%
```