@oflebbe 2021-04-02
Ok this will introduce a huge Garbage Collection overhead, since we will have 8000x more objects. 
My impression the library will be used either to take all objects from a result or to copy away the few needed. 
I totally see the issue with #25  , where I proposed a way to use like I am using consuming the lib.

We cannot use sync.Pool as well, since the consumer of the library who has to do Put() after discarding a result.

I am thinking about having a different api, where the consumer of the api has to call the Scanner with a function as a predicate to find whether a given node/way/relation should be returned or not. What do you think @paulmach  ?

Or even doing the complete processing by a user-supplied function (as I have seen it in a rust osm library).