@paulmach 2021-04-27
After thinking about this, I think this change is necessary to support the current API.  Then we can add something like https://github.com/paulmach/osm/pull/27 for anyone that needs the extra performance.

The way I'm thinking about this is objects in the `for scanner.Scan() {}` loop can have 3ish different lifetimes:
1. Just within the loop. Maybe you're counting or transforming and saving, or just inserting into a DB.
2. For a few loops. Like bulk inserting into the DB.
3. Forever, you're saving a handful of the nodes into a map.

1+2 work today, 3 creates high memory usages as reported in https://github.com/paulmach/osm/issues/25. To support all 3 cases with the current API we need this PR.

Performance wise I think we're okay. Comparing v0.1.1 (the release that added this to be reverted change) to this branch I see:
```
benchmark              old ns/op     new ns/op     delta
BenchmarkLondon-12     355943390     251720242     -29.28%

benchmark              old allocs     new allocs     delta
BenchmarkLondon-12     7732810        5145480        -33.46%

benchmark              old bytes      new bytes     delta
BenchmarkLondon-12     1306695743     952781394     -27.08%
```
So we're still better than what we had when the change was originally proposed.